♻️ FX58BG task: split init orchestrator helpers
